account auth/creation/validation/password reset

+++ src/components/authentication/Login.js
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 // Axios instance
-import instance from "../axios/axios";
+import instance from "../../axios/axios";
 // Login - CSS
 import "./Login.scss";
 
-function Login({ setIsAuth }) {
+function Login() {
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
 
@@ -18,12 +18,16 @@ function Login({ setIsAuth }) {
     // create POST request
     const { data } = await instance.post("authentication/token/", user);
 
-    // initialize access and refresh tokens in localstorage
-    localStorage.clear();
-    localStorage.setItem("access_token", data.access);
-    localStorage.setItem("refresh_token", data.refresh);
-    // redirect
-    window.location.href = "/dashboard";
+    if (data) {
+      // initialize access and refresh tokens in localstorage
+      localStorage.clear();
+      localStorage.setItem("access_token", data.access);
+      localStorage.setItem("refresh_token", data.refresh);
+      // redirect
+      window.location.href = "/dashboard";
+    } else {
+      console.log("login error");
+    }
   };
 
   return (
@@ -35,7 +39,7 @@ function Login({ setIsAuth }) {
         value={username}
         onChange={(e) => setUsername(e.target.value)}
         required></input>
-      <label htmlFor="username">password:</label>
+      <label htmlFor="password">password:</label>
       <input
         name="password"
         type="password"
